created abstract base classes for *Period-models
Functionality of those 2 models are now abstract and can be used in more ways
than before. The abstract classes don't depend on PeriodModel (which has a
generic-relation) and a custom PeriodModel with specific relations might be
implemented.
